+++ packages/backend-ai/src/index.ts
@@ -4,7 +4,16 @@ import {
   ConversationHandlerResources,
 } from './conversation/conversation_handler_construct.js';
 
+import { ConversationTurnEvent } from './conversation/lambda/types.js';
+
+import { ConversationTurnExecutor } from './conversation/lambda/conversation_turn_executor.js';
+
+import { ConversationTurnResponder } from './conversation/lambda/conversation_turn_responder.js';
+
 export {
+  ConversationTurnEvent,
+  ConversationTurnExecutor,
+  ConversationTurnResponder,
   ConversationHandler,
   ConversationHandlerProps,
   ConversationHandlerResources,
