+++ test-projects/conversation-test/amplify/backend.ts
@@ -1,6 +1,7 @@
 import { defineBackend } from '@aws-amplify/backend';
 import { auth } from './auth/resource';
 import { data } from './data/resource';
+import { evilChatHandler } from './evilChatHandler/resource';
 
 /**
  * @see https://docs.amplify.aws/react/build-a-backend/ to add storage, functions, and more
@@ -8,4 +9,5 @@ import { data } from './data/resource';
 const backend = defineBackend({
   auth,
   data,
+  evilChatHandler,
 });
