+++ test-projects/conversation-test/amplify/data/resource.ts
@@ -11,6 +11,11 @@ const schema = `
         aiModel: "Claude3Haiku",
         systemPrompt: "You are a helpful chatbot. Respond in 20 words or less."
       )
+      
+      evilChat(sessionId: ID, content: String): String
+      @conversation(
+         functionName: "amplify-conversationtest--evilChatHandlerlambdaC10-b1tKw5p9t5bE"
+      )
     }
 `;
 
@@ -21,21 +26,3 @@ export const data = defineData({
   },
 });
 
-// Previous schemas.
-
-/*
-
-    type Foo @model(subscriptions: { level: off }, mutations: null, queries: null) {
-      bar: Int
-    }
-
-    type Mutation {
-      pirateChat(sessionId: ID, content: String): String
-      @conversation(
-        aiModel: "Claude3Haiku",
-        functionName: "amplify-conversationtest--conversationapp2lambdaha-jka2V1S0O98j",
-            systemPrompt: "You are a helpful chatbot. Respond in 20 words or less."
-      )
-    }
-
- */
