+++ test-projects/conversation-test/amplify/evilChatHandler/resource.ts
@@ -1,0 +1,3 @@
+import { defineFunction } from '@aws-amplify/backend';
+
+export const evilChatHandler = defineFunction();
